added some small UI changes and refactored many of the class names to be more clear and consistant. also changed app title.
