+++ MyApplication/app/src/main/res/layout/activity_game_with_gold_details.xml
@@ -13,7 +13,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.colinbradley.xboxoneutilitiesapp.store_page.games_with_gold.GameWithGoldDetails">
+    tools:context="com.colinbradley.xboxoneutilitiesapp.store_page.games_with_gold.GWGDetails">
 
     <ImageView
         android:id="@+id/gwgd_main_img"
