+++ MyApplication/app/src/main/res/layout/activity_profile.xml
@@ -109,11 +109,6 @@
         android:textColor="@color/textColor"
         />
 
-    <ProgressBar
-        android:id="@+id/profile_progressbar"
-        android:layout_width="50sp"
-        android:layout_height="50sp"
-        android:layout_centerInParent="true"/>
 
         <android.support.v7.widget.Toolbar
             android:id="@+id/profile_toolbar"
@@ -132,6 +127,7 @@
                 app:tabTextColor="@color/textColor"
                 app:tabIndicatorColor="@color/colorAccent"
                 app:tabSelectedTextColor="@color/colorAccent"
+                android:layout_marginRight="12sp"
                 />
 
             </android.support.v7.widget.Toolbar>
