+++ MyApplication/app/src/main/res/layout/activity_store.xml
@@ -26,7 +26,9 @@
             app:tabSelectedTextColor="@color/colorAccent"
             android:layout_alignParentTop="true"
             android:layout_alignParentRight="true"
-            android:layout_alignParentEnd="true" />
+            android:layout_alignParentEnd="true"
+            android:layout_marginRight="12sp"
+            />
     </android.support.v7.widget.Toolbar>
