+++ MyApplication/app/src/main/res/layout/fragment_gameclips.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical"
     android:layout_width="match_parent"
@@ -10,4 +10,10 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
 
-</LinearLayout>
+    <ProgressBar
+        android:id="@+id/gameclips_progressbar"
+        android:layout_width="60sp"
+        android:layout_height="60sp"
+        android:layout_centerInParent="true"/>
+
+</RelativeLayout>
